Trait reth_interfaces::blockchain_tree::BlockchainTreeViewer

pub trait BlockchainTreeViewer: Send + Sync {

    // Required methods
    fn blocks(&self) -> BTreeMap<BlockNumber, HashSet<BlockHash>>;
    fn header_by_hash(&self, hash: BlockHash) -> Option<SealedHeader>;
    fn block_by_hash(&self, hash: BlockHash) -> Option<SealedBlock>;
    fn block_with_senders_by_hash(
        &self,
        hash: BlockHash
    ) -> Option<SealedBlockWithSenders>;
    fn buffered_block_by_hash(
        &self,
        block_hash: BlockHash
    ) -> Option<SealedBlock>;
    fn buffered_header_by_hash(
        &self,
        block_hash: BlockHash
    ) -> Option<SealedHeader>;
    fn canonical_blocks(&self) -> BTreeMap<BlockNumber, BlockHash>;
    fn find_canonical_ancestor(
        &self,
        parent_hash: BlockHash
    ) -> Option<BlockHash>;
    fn is_canonical(&self, hash: BlockHash) -> RethResult<bool>;
    fn lowest_buffered_ancestor(
        &self,
        hash: BlockHash
    ) -> Option<SealedBlockWithSenders>;
    fn canonical_tip(&self) -> BlockNumHash;
    fn pending_blocks(&self) -> (BlockNumber, Vec<BlockHash>);
    fn pending_block_num_hash(&self) -> Option<BlockNumHash>;
    fn pending_block_and_receipts(&self) -> Option<(SealedBlock, Vec<Receipt>)>;
    fn receipts_by_block_hash(
        &self,
        block_hash: BlockHash
    ) -> Option<Vec<Receipt>>;

    // Provided methods
    fn contains(&self, hash: BlockHash) -> bool { ... }
    fn pending_block(&self) -> Option<SealedBlock> { ... }
    fn pending_block_with_senders(&self) -> Option<SealedBlockWithSenders> { ... }
    fn pending_receipts(&self) -> Option<Vec<Receipt>> { ... }
    fn pending_header(&self) -> Option<SealedHeader> { ... }

}

Allows read only functionality on the blockchain tree.

Tree contains all blocks that are not canonical that can potentially be included as canonical chain. For better explanation we can group blocks into four groups:

Canonical chain blocks

Side chain blocks. Side chain are block that forks from canonical chain but not its tip.

Pending blocks that extend the canonical chain but are not yet included.

Future pending blocks that extend the pending blocks.

Required Methods

fn blocks(&self) -> BTreeMap<BlockNumber, HashSet<BlockHash>>

Returns both pending and side-chain block numbers and their hashes.

Caution: This will not return blocks from the canonical chain.

fn header_by_hash(&self, hash: BlockHash) -> Option<SealedHeader>

Returns the header with matching hash from the tree, if it exists.

Caution: This will not return headers from the canonical chain.

fn block_by_hash(&self, hash: BlockHash) -> Option<SealedBlock>

Returns the block with matching hash from the tree, if it exists.

Caution: This will not return blocks from the canonical chain or buffered blocks that are disconnected from the canonical chain.

fn block_with_senders_by_hash( &self, hash: BlockHash ) -> Option<SealedBlockWithSenders>

Returns the block with matching hash from the tree, if it exists.

Caution: This will not return blocks from the canonical chain or buffered blocks that are disconnected from the canonical chain.

fn buffered_block_by_hash(&self, block_hash: BlockHash) -> Option<SealedBlock>

Returns the buffered (disconnected) block with matching hash from the internal buffer if it exists.

Caution: Unlike Self::block_by_hash this will only return blocks that are currently disconnected from the canonical chain.

fn buffered_header_by_hash(&self, block_hash: BlockHash) -> Option<SealedHeader>

Returns the buffered (disconnected) header with matching hash from the internal buffer if it exists.

Caution: Unlike Self::block_by_hash this will only return headers that are currently disconnected from the canonical chain.

fn canonical_blocks(&self) -> BTreeMap<BlockNumber, BlockHash>

Canonical block number and hashes best known by the tree.

fn find_canonical_ancestor(&self, parent_hash: BlockHash) -> Option<BlockHash>

Given the parent hash of a block, this tries to find the last ancestor that is part of the canonical chain.

In other words, this will walk up the (side) chain starting with the given hash and return the first block that’s canonical.

Note: this could be the given parent_hash if it’s already canonical.

fn is_canonical(&self, hash: BlockHash) -> RethResult<bool>

Return whether or not the block is known and in the canonical chain.

fn lowest_buffered_ancestor( &self, hash: BlockHash ) -> Option<SealedBlockWithSenders>

Given the hash of a block, this checks the buffered blocks for the lowest ancestor in the buffer.

If there is a buffered block with the given hash, this returns the block itself.

fn canonical_tip(&self) -> BlockNumHash

Return BlockchainTree best known canonical chain tip (BlockHash, BlockNumber)

fn pending_blocks(&self) -> (BlockNumber, Vec<BlockHash>)

Return block hashes that extends the canonical chain tip by one. This is used to fetch what is considered the pending blocks, blocks that has best chance to become canonical.

fn pending_block_num_hash(&self) -> Option<BlockNumHash>

Return block number and hash that extends the canonical chain tip by one.

If there is no such block, this returns None.

fn pending_block_and_receipts(&self) -> Option<(SealedBlock, Vec<Receipt>)>

Returns the pending block and its receipts in one call.

This exists to prevent a potential data race if the pending block changes in between Self::pending_block and Self::pending_receipts calls.

fn receipts_by_block_hash(&self, block_hash: BlockHash) -> Option<Vec<Receipt>>

Returns the pending receipts if there is one.

Provided Methods

fn contains(&self, hash: BlockHash) -> bool

Returns true if the tree contains the block with matching hash.

fn pending_block(&self) -> Option<SealedBlock>

Returns the pending block if there is one.

fn pending_block_with_senders(&self) -> Option<SealedBlockWithSenders>

Returns the pending block if there is one.

fn pending_receipts(&self) -> Option<Vec<Receipt>>

Returns the pending receipts if there is one.

fn pending_header(&self) -> Option<SealedHeader>

Returns the pending block if there is one.