Module reth_interfaces::p2p

P2P traits.

Modules

bodies

Traits for implementing P2P block body clients.

download

Shared abstractions for downloader implementations.

either

A downloader that combines two different downloaders/client implementations.

error

Error types broadly used by p2p interfaces for any operation which may produce an error when interacting with the network implementation

full_block

An implementation that uses headers and bodies traits to download full blocks

headers

Traits for implementing P2P Header Clients. Also includes implementations of a Linear and a Parallel downloader generic over the Consensus and HeadersClient.

priority

Priority enum for BlockHeader and BlockBody requests