Trait reth_interfaces::p2p::bodies::client::BodiesClient

pub trait BodiesClient: DownloadClient {
    type Output: Future<Output = PeerRequestResult<Vec<BlockBody>>> + Sync + Send + Unpin;

    // Required method
    fn get_block_bodies_with_priority(
        &self,
        hashes: Vec<B256>,
        priority: Priority
    ) -> Self::Output;

    // Provided methods
    fn get_block_bodies(&self, hashes: Vec<B256>) -> Self::Output { ... }
    fn get_block_body(&self, hash: B256) -> SingleBodyRequest<Self::Output>  { ... }
    fn get_block_body_with_priority(
        &self,
        hash: B256,
        priority: Priority
    ) -> SingleBodyRequest<Self::Output>  { ... }
}

A client capable of downloading block bodies.

Required Associated Types

type Output: Future<Output = PeerRequestResult<Vec<BlockBody>>> + Sync + Send + Unpin

The output of the request future for querying block bodies.

Required Methods

fn get_block_bodies_with_priority( &self, hashes: Vec<B256>, priority: Priority ) -> Self::Output

Fetches the block body for the requested block with priority

Provided Methods

fn get_block_bodies(&self, hashes: Vec<B256>) -> Self::Output

Fetches the block body for the requested block.

fn get_block_body(&self, hash: B256) -> SingleBodyRequest<Self::Output>

Fetches a single block body for the requested hash.

fn get_block_body_with_priority( &self, hash: B256, priority: Priority ) -> SingleBodyRequest<Self::Output>

Fetches a single block body for the requested hash with priority