Trait reth_interfaces::p2p::bodies::downloader::BodyDownloader

pub trait BodyDownloader: Send + Sync + Stream<Item = BodyDownloaderResult> + Unpin {
    // Required method
    fn set_download_range(
        &mut self,
        range: RangeInclusive<BlockNumber>
    ) -> DownloadResult<()>;
}

A downloader capable of fetching and yielding block bodies from block headers.

A downloader represents a distinct strategy for submitting requests to download block bodies, while a BodiesClient represents a client capable of fulfilling these requests.

Required Methods

fn set_download_range( &mut self, range: RangeInclusive<BlockNumber> ) -> DownloadResult<()>

Method for setting the download range.