Module reth_interfaces::p2p::error

Error types broadly used by p2p interfaces for any operation which may produce an error when interacting with the network implementation

Enums

DownloadError

The downloader error type

RequestError

Error variants that can happen when sending requests to a session.

Traits

EthResponseValidator

Helper trait used to validate responses.

Type Aliases

DownloadResult

The download result type

PeerRequestResult

Result with PeerId

RequestResult

Result alias for result of a request.