Trait reth_interfaces::p2p::headers::client::HeadersClient

pub trait HeadersClient: DownloadClient {
    type Output: Future<Output = PeerRequestResult<Vec<Header>>> + Sync + Send + Unpin;

    // Required method
    fn get_headers_with_priority(
        &self,
        request: HeadersRequest,
        priority: Priority
    ) -> Self::Output;

    // Provided methods
    fn get_headers(&self, request: HeadersRequest) -> Self::Output { ... }
    fn get_header(
        &self,
        start: BlockHashOrNumber
    ) -> SingleHeaderRequest<Self::Output>  { ... }
    fn get_header_with_priority(
        &self,
        start: BlockHashOrNumber,
        priority: Priority
    ) -> SingleHeaderRequest<Self::Output>  { ... }
}

The block headers downloader client

Required Associated Types

type Output: Future<Output = PeerRequestResult<Vec<Header>>> + Sync + Send + Unpin

The headers future type

Required Methods

fn get_headers_with_priority( &self, request: HeadersRequest, priority: Priority ) -> Self::Output

Sends the header request to the p2p network with priority set and returns the header response received from a peer.

Provided Methods

fn get_headers(&self, request: HeadersRequest) -> Self::Output

Sends the header request to the p2p network and returns the header response received from a peer.

fn get_header( &self, start: BlockHashOrNumber ) -> SingleHeaderRequest<Self::Output>

Fetches a single header for the requested number or hash.

fn get_header_with_priority( &self, start: BlockHashOrNumber, priority: Priority ) -> SingleHeaderRequest<Self::Output>

Fetches a single header for the requested number or hash with priority