Module reth_interfaces::test_utils

Common test helpers for mocking out Consensus, Downloaders and Header Clients.

Modules

generators

Generators for different data structures like block headers, block bodies and ranges of those.

Structs

NoopFullBlockClient

A headers+bodies client implementation that does nothing.

TestBodiesClient

A test client for fetching bodies

TestConsensus

Consensus engine implementation for testing

TestFullBlockClient

A headers+bodies client that stores the headers and bodies in memory, with an artificial soft bodies response limit that is set to 20 by default.

TestHeaderDownloader

A test downloader which just returns the values that have been pushed to it.

TestHeadersClient

A test client for fetching headers