Module reth_interfaces::test_utils::generators

Generators for different data structures like block headers, block bodies and ranges of those.

Traits

Rng

An automatically-implemented extension trait on RngCore providing high-level generic methods for sampling values and other convenience methods.

Functions

generate_keys

Generates a set of KeyPairs based on the desired count.

random_account_change

Generate a random account change.

random_block

Generate a random block filled with signed transactions (generated using random_signed_tx). If no transaction count is provided, the number of transactions will be random, otherwise the provided count will be used.

random_block_range

Generate a range of random blocks.

random_changeset_range

Generate a range of changesets for given blocks and accounts. Assumes all accounts start with an empty storage.

random_contract_account_range

Generate random Contract Accounts

random_eoa_account

Generate random Externally Owned Account (EOA account without contract).

random_eoa_account_range

Generate random Externally Owned Accounts

random_header

Generate a random SealedHeader.

random_header_range

Generates a range of random SealedHeaders.

random_log

Generate random log

random_receipt

Generate random receipt for transaction

random_signed_tx

Generates a random legacy Transaction that is signed.

random_storage_entry

Generate a random storage change.

random_tx

Generates a random legacy Transaction.

rng

Returns a random number generator that can be seeded using the SEED environment variable.

sign_tx_with_key_pair

Signs the Transaction with the given key pair.

Type Aliases

ChangeSet

Collection of account and storage entry changes