Crate reth_interfaces

A collection of shared traits and error types used in Reth.

Feature Flags

test-utils: Export utilities for testing

Modules

blockchain_tree

BlockchainTree related traits.

consensus

Consensus traits.

db

Database error

executor

Block Execution traits.

p2p

P2P traits.

provider

Provider error

sync

Syncing related traits. Traits used when interacting with the sync status of the network.

Enums

RethError

Core error variants possible when interacting with the blockchain.

Type Aliases

RethResult

Result alias for RethError.